pyproject.toml
[project]
name = "powershell-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "openai>=1.78.1",
    "pyinstaller>=6.11.1",
    "pyperclip>=1.9.0",
    "python-dotenv>=1.0.1",
]

[tool.setuptools.packages.find]
where = ["src"]
